# src/main/resources/static/openapi/profile-settings.yaml
openapi: 3.1.0
info:
  title: Broke.AI Profile and Settings API
  version: 1.0.0
  description: |
    Production contract for signed-in profile, security, preferences, privacy,
    data controls, and support features. App-lock secrets and biometric data are
    intentionally excluded and must remain protected on the user's device.

    Password implementations must accept long passphrases without arbitrary
    composition rules, reject compromised/common passwords, rate-limit sensitive
    operations, and hash new passwords with Argon2id. Existing bcrypt hashes may
    be upgraded after successful authentication.
servers:
  - url: /api
    description: Application API base path
security:
  - bearerAuth: []
...
paths:
...
  /v1/guest/transactions/clear:
    post:
      tags: [Guest account]
      operationId: clearGuestTransactionHistory
      summary: Soft-delete every active transaction owned by the current guest
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestClearRequest'
      responses:
        '200':
          description: Guest transaction history cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClearTransactionsEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/guest/data-summary:
    get:
      tags: [Guest account]
      operationId: getGuestDataSummary
      summary: Get guest transaction and AI allowance totals without loading history
      responses:
        '200':
          description: Guest data summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestDataSummaryEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/guest-account:
    delete:
      tags: [Guest account]
      operationId: deleteGuestAccount
      summary: Permanently remove guest-owned data and revoke the guest identity
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestDeleteRequest'
      responses:
        '200':
          description: Guest data removed; the client may now clear its local session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestDeletionEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 255
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: VALIDATION_ERROR
              message: The request contains invalid data.
              field: fullName
              requestId: 550e8400-e29b-41d4-a716-446655440000
    Unauthenticated:
      description: Authentication is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHENTICATED
              message: Authentication is required.
              requestId: 550e8400-e29b-41d4-a716-446655440000
    Forbidden:
      description: The authenticated account type is not allowed to perform this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: Sensitive-operation rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until another attempt is allowed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: RATE_LIMITED
              message: Too many attempts. Try again later.
              requestId: 550e8400-e29b-41d4-a716-446655440000
...
  schemas:
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [requestId, serverTime]
      properties:
        requestId:
          type: string
          format: uuid
        serverTime:
          type: string
          format: date-time
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - INVALID_CREDENTIALS
        - GUEST_ONLY
        - REAUTHENTICATION_REQUIRED
        - CURRENT_PASSWORD_INVALID
        - PASSWORD_COMPROMISED
        - USERNAME_ALREADY_USED
        - EMAIL_ALREADY_USED
        - EMAIL_VERIFICATION_REQUIRED
        - RATE_LIMITED
        - CONFLICTING_UPDATE
        - EXPORT_NOT_READY
        - RESOURCE_NOT_FOUND
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            field:
              type: [string, 'null']
            requestId:
              type: string
              format: uuid
...
    GuestClearRequest:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation:
          type: string
          const: CLEAR
    GuestDeleteRequest:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation:
          type: string
          const: DELETE
    GuestDeletionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [deleted]
          properties:
            deleted:
              type: boolean
              const: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    GuestDataSummaryEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [transactionCount, remainingAiTrials]
          properties:
            transactionCount:
              type: integer
              format: int64
              minimum: 0
            firstTransactionAt:
              type: [string, 'null']
              format: date-time
            lastTransactionAt:
              type: [string, 'null']
              format: date-time
            remainingAiTrials:
              type: integer
              minimum: 0
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ClearTransactionsEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [deletedCount, recoverableUntil]
          properties:
            deletedCount:
              type: integer
              format: int64
              minimum: 0
            recoverableUntil:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...